Format library version independently of *READ-DEFAULT-FLOAT-FORMAT*
Before this change, the format control string ~A was used to format
the library version stored in *VERSION*. Since that variable is
defined as

  (defparameter *version* 2.03)

the value is either a SINGLE-FLOAT or a DOUBLE-FLOAT, depending on the
value of *READ-DEFAULT-FLOAT-FORMAT* at the time the cl-pdf source
code is read. In case the value of *READ-DEFAULT-FLOAT-FORMAT* at
runtime differs from the earlier value, ~A prints a float exponent
marker ("f" or "d") which is confusing.